Fix defect with bot users when 'delete_previous_comments' is enabled
Indeed, in order to delete the correct comments, the golang code makes sure
that comment's `Author.Login` is the same as current `Viewer.Login`. But with
bot users, viewer's login is "githubApp[bot]" and author login is "githubApp".

As a consequence, for the comparison to succeed, we need to remove any
trailing "[bot]" suffix from both parties, and compare the resulting prefixes.

This isue was originally submitted as telia-oss/github-pr-resource#245 and
fixed later in monzo/github-pr-resource#4.